Move 'koodi' helper functions to a utils/koodi subdirectory
