+++ src/main/app/src/components/AsyncKoodistoSelect/index.jsx
@@ -6,7 +6,7 @@ import HttpContext from '../HttpContext';
 import UrlContext from '../UrlContext';
 import useLanguage from '../useLanguage';
 import getKoodiNimiTranslation from '../../utils/getKoodiNimiTranslation';
-import parseKoodiUri from '../../utils/parseKoodiUri';
+import parseKoodiUri from '#/src/utils/koodi/parseKoodiUri';
 import { isFunction } from '../../utils';
 
 const AsyncKoodistoSelect = ({
